* jruby-1_7: (54 commits)
  [travis-ci] re-try sudo: false
  dry out the two (Java helper) respond_to_p19 methods
  respondsTo JRuby API under 1.9.3 does not do respond_to_missing? like MRI does
  sleep should convert it's "time interval" argument the same way as MRI
  respond_to_missing? should always get a symbol name passed (fixes #2946)
  add a respond_to_missing unit test (on 1.9)
  introduce and use conversion( just like MRI does) on #timeout's sec argument
  add missing ' ' (space) in exception message ... TypeError from `sleep('invalid')`
  DateTime.iso8601 fails with an error if a second fraction is present
  [build] switch to 1.7.21-SNAPSHOT for test-jruby-jars
  refactor Timeout's scheduled tasks to an inner class (for easier identification)
  further hand drying of the Timeout module impl
  less (int) casting noiz
  add some (internal) generics for an improved readability xp
  inner class' Finalizer fields id and finalized can be final
  generix-ize the environment field (at least internally) + use a StringBuilder
  unnecessary (int) cast
  missing throw statement in RubyString's concat
  tune timeout to work when scheduling or argument convenrsion throws (see #2940)
  return empty Ruby array for empty dir list
  ...

Conflicts:
	.gitignore
	.travis.yml
	VERSION
	core/pom.xml
	core/src/main/java/org/jruby/RubyArray.java
	core/src/main/java/org/jruby/RubyBasicObject.java
	core/src/main/java/org/jruby/RubyDir.java
	core/src/main/java/org/jruby/RubyEnumerable.java
	core/src/main/java/org/jruby/RubyFile.java
	core/src/main/java/org/jruby/RubyGlobal.java
	core/src/main/java/org/jruby/RubyIO.java
	core/src/main/java/org/jruby/RubyKernel.java
	core/src/main/java/org/jruby/RubyModule.java
	core/src/main/java/org/jruby/RubyString.java
	core/src/main/java/org/jruby/RubyTime.java
	core/src/main/java/org/jruby/ext/timeout/Timeout.java
	core/src/main/java/org/jruby/java/proxies/JavaProxy.java
	core/src/main/java/org/jruby/util/Dir.java
	core/src/main/java/org/jruby/util/encoding/CharsetTranscoder.java
	docs/man/pom.xml
	docs/pom.xml
	ext/pom.xml
	ext/readline/pom.xml
	ext/ripper/pom.xml
	lib/pom.xml
	lib/ruby/1.9/date/format.rb
	lib/ruby/1.9/io/linux_console.rb
	lib/ruby/shared/gauntlet_rubygems.rb
	lib/ruby/shared/rubygems.rb
	lib/ruby/shared/rubygems/available_set.rb
	lib/ruby/shared/rubygems/basic_specification.rb
	lib/ruby/shared/rubygems/command.rb
	lib/ruby/shared/rubygems/command_manager.rb
	lib/ruby/shared/rubygems/commands/cert_command.rb
	lib/ruby/shared/rubygems/commands/cleanup_command.rb
	lib/ruby/shared/rubygems/commands/contents_command.rb
	lib/ruby/shared/rubygems/commands/dependency_command.rb
	lib/ruby/shared/rubygems/commands/environment_command.rb
	lib/ruby/shared/rubygems/commands/generate_index_command.rb
	lib/ruby/shared/rubygems/commands/help_command.rb
	lib/ruby/shared/rubygems/commands/install_command.rb
	lib/ruby/shared/rubygems/commands/list_command.rb
	lib/ruby/shared/rubygems/commands/mirror_command.rb
	lib/ruby/shared/rubygems/commands/outdated_command.rb
	lib/ruby/shared/rubygems/commands/owner_command.rb
	lib/ruby/shared/rubygems/commands/pristine_command.rb
	lib/ruby/shared/rubygems/commands/push_command.rb
	lib/ruby/shared/rubygems/commands/query_command.rb
	lib/ruby/shared/rubygems/commands/search_command.rb
	lib/ruby/shared/rubygems/commands/setup_command.rb
	lib/ruby/shared/rubygems/commands/specification_command.rb
	lib/ruby/shared/rubygems/commands/uninstall_command.rb
	lib/ruby/shared/rubygems/commands/unpack_command.rb
	lib/ruby/shared/rubygems/commands/update_command.rb
	lib/ruby/shared/rubygems/commands/which_command.rb
	lib/ruby/shared/rubygems/commands/yank_command.rb
	lib/ruby/shared/rubygems/compatibility.rb
	lib/ruby/shared/rubygems/config_file.rb
	lib/ruby/shared/rubygems/core_ext/kernel_gem.rb
	lib/ruby/shared/rubygems/core_ext/kernel_require.rb
	lib/ruby/shared/rubygems/defaults.rb
	lib/ruby/shared/rubygems/dependency.rb
	lib/ruby/shared/rubygems/dependency_installer.rb
	lib/ruby/shared/rubygems/dependency_list.rb
	lib/ruby/shared/rubygems/deprecate.rb
	lib/ruby/shared/rubygems/doctor.rb
	lib/ruby/shared/rubygems/errors.rb
	lib/ruby/shared/rubygems/exceptions.rb
	lib/ruby/shared/rubygems/ext.rb
	lib/ruby/shared/rubygems/ext/builder.rb
	lib/ruby/shared/rubygems/ext/cmake_builder.rb
	lib/ruby/shared/rubygems/ext/configure_builder.rb
	lib/ruby/shared/rubygems/ext/ext_conf_builder.rb
	lib/ruby/shared/rubygems/ext/rake_builder.rb
	lib/ruby/shared/rubygems/gemcutter_utilities.rb
	lib/ruby/shared/rubygems/indexer.rb
	lib/ruby/shared/rubygems/install_update_options.rb
	lib/ruby/shared/rubygems/installer.rb
	lib/ruby/shared/rubygems/installer_test_case.rb
	lib/ruby/shared/rubygems/local_remote_options.rb
	lib/ruby/shared/rubygems/name_tuple.rb
	lib/ruby/shared/rubygems/package.rb
	lib/ruby/shared/rubygems/package/old.rb
	lib/ruby/shared/rubygems/package/tar_header.rb
	lib/ruby/shared/rubygems/package/tar_reader/entry.rb
	lib/ruby/shared/rubygems/package/tar_writer.rb
	lib/ruby/shared/rubygems/path_support.rb
	lib/ruby/shared/rubygems/platform.rb
	lib/ruby/shared/rubygems/psych_additions.rb
	lib/ruby/shared/rubygems/rdoc.rb
	lib/ruby/shared/rubygems/remote_fetcher.rb
	lib/ruby/shared/rubygems/request.rb
	lib/ruby/shared/rubygems/request_set.rb
	lib/ruby/shared/rubygems/request_set/gem_dependency_api.rb
	lib/ruby/shared/rubygems/requirement.rb
	lib/ruby/shared/rubygems/resolver/current_set.rb
	lib/ruby/shared/rubygems/security.rb
	lib/ruby/shared/rubygems/security/policy.rb
	lib/ruby/shared/rubygems/security/signer.rb
	lib/ruby/shared/rubygems/security/trust_dir.rb
	lib/ruby/shared/rubygems/server.rb
	lib/ruby/shared/rubygems/source.rb
	lib/ruby/shared/rubygems/source/installed.rb
	lib/ruby/shared/rubygems/source/local.rb
	lib/ruby/shared/rubygems/source/specific_file.rb
	lib/ruby/shared/rubygems/source_list.rb
	lib/ruby/shared/rubygems/spec_fetcher.rb
	lib/ruby/shared/rubygems/specification.rb
	lib/ruby/shared/rubygems/ssl_certs/AddTrustExternalCARoot.pem
	lib/ruby/shared/rubygems/stub_specification.rb
	lib/ruby/shared/rubygems/syck_hack.rb
	lib/ruby/shared/rubygems/test_case.rb
	lib/ruby/shared/rubygems/test_utilities.rb
	lib/ruby/shared/rubygems/text.rb
	lib/ruby/shared/rubygems/uninstaller.rb
	lib/ruby/shared/rubygems/uri_formatter.rb
	lib/ruby/shared/rubygems/user_interaction.rb
	lib/ruby/shared/rubygems/util/list.rb
	lib/ruby/shared/rubygems/validator.rb
	lib/ruby/shared/rubygems/version.rb
	maven/jruby-complete/pom.xml
	maven/jruby-dist/pom.xml
	maven/jruby-jars/pom.xml
	maven/jruby-noasm/pom.xml
	maven/jruby-rake-plugin/pom.xml
	maven/jruby-stdlib/pom.xml
	maven/jruby/pom.xml
	maven/pom.xml
	pom.xml
	test/pom.xml
	test/test_dir.rb
	test/test_kernel.rb
	test/test_respond_to.rb
	test/test_timeout.rb

+++ core/src/main/java/org/jruby/RubyInteger.java
@@ -19,7 +19,7 @@
  * Copyright (C) 2002-2004 Thomas E Enebo <enebo@acm.org>
  * Copyright (C) 2004 Stefan Matthias Aust <sma@3plus4.de>
  * Copyright (C) 2005 Charles O Nutter <headius@headius.com>
- * 
+ *
  * Alternatively, the contents of this file may be used under the terms of
  * either of the GNU General Public License Version 2 or later (the "GPL"),
  * or the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
@@ -62,7 +62,7 @@ import static org.jruby.RubyEnumerator.SizeFn;
  * @author  jpetersen
  */
 @JRubyClass(name="Integer", parent="Numeric", include="Precision")
-public abstract class RubyInteger extends RubyNumeric { 
+public abstract class RubyInteger extends RubyNumeric {
 
     public static RubyClass createIntegerClass(Ruby runtime) {
         RubyClass integer = runtime.defineClass("Integer", runtime.getNumeric(),
@@ -71,13 +71,13 @@ public abstract class RubyInteger extends RubyNumeric {
 
         integer.setClassIndex(ClassIndex.INTEGER);
         integer.setReifiedClass(RubyInteger.class);
-        
+
         integer.kindOf = new RubyModule.JavaClassKindOf(RubyInteger.class);
 
         integer.getSingletonClass().undefineMethod("new");
 
         integer.defineAnnotatedMethods(RubyInteger.class);
-        
+
         return integer;
     }
 
@@ -88,15 +88,15 @@ public abstract class RubyInteger extends RubyNumeric {
     public RubyInteger(RubyClass rubyClass) {
         super(rubyClass);
     }
-    
+
     public RubyInteger(Ruby runtime, RubyClass rubyClass, boolean useObjectSpace) {
         super(runtime, rubyClass, useObjectSpace);
-    }     
+    }
 
     @Deprecated
     public RubyInteger(Ruby runtime, RubyClass rubyClass, boolean useObjectSpace, boolean canBeTainted) {
         super(runtime, rubyClass, useObjectSpace, canBeTainted);
-    }     
+    }
 
     @Override
     public RubyInteger convertToInteger() {
@@ -110,11 +110,11 @@ public abstract class RubyInteger extends RubyNumeric {
 
     /*  ================
      *  Instance Methods
-     *  ================ 
+     *  ================
      */
 
     /** int_int_p
-     * 
+     *
      */
     @Override
     @JRubyMethod(name = "integer?")
@@ -123,7 +123,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** int_upto
-     * 
+     *
      */
     @JRubyMethod
     public IRubyObject upto(ThreadContext context, IRubyObject to, Block block) {
@@ -185,7 +185,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** int_downto
-     * 
+     *
      */
     // TODO: Make callCoerced work in block context...then fix downto, step, and upto.
     @JRubyMethod
@@ -283,7 +283,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** int_succ
-     * 
+     *
      */
     @JRubyMethod(name = {"succ", "next"})
     public IRubyObject succ(ThreadContext context) {
@@ -312,7 +312,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** int_chr
-     * 
+     *
      */
     public RubyString chr(ThreadContext context) {
         return chr19(context);
@@ -331,7 +331,7 @@ public abstract class RubyInteger extends RubyNumeric {
                 throw runtime.newRangeError(this.toString() + " out of char range");
             } else {
                 if (enc == null) enc = USASCIIEncoding.INSTANCE;
-                return RubyString.newStringNoCopy(runtime, fromEncodedBytes(runtime, enc, (int)value), enc, 0);
+                return RubyString.newStringNoCopy(runtime, fromEncodedBytes(runtime, enc, value), enc, 0);
             }
         }
     }
@@ -361,7 +361,7 @@ public abstract class RubyInteger extends RubyNumeric {
         }
 
         if (n <= 0) throw runtime.newRangeError(this.toString() + " out of char range");
-        
+
         ByteList bytes = new ByteList(n);
 
         boolean ok = false;
@@ -381,7 +381,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** int_ord
-     * 
+     *
      */
     @JRubyMethod(name = "ord")
     public IRubyObject ord(ThreadContext context) {
@@ -389,7 +389,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** int_to_i
-     * 
+     *
      */
     @JRubyMethod(name = {"to_i", "to_int", "floor", "ceil", "truncate"})
     public IRubyObject to_i() {
@@ -412,14 +412,14 @@ public abstract class RubyInteger extends RubyNumeric {
         if (ndigits > 0) return RubyKernel.new_float(this, this);
         if (ndigits == 0) return this;
         Ruby runtime = context.runtime;
-        
+
         long bytes = (this instanceof RubyFixnum) ? 8 : RubyFixnum.fix2long(callMethod("size"));
         /* If 10**N/2 > this, return 0 */
         /* We have log_256(10) > 0.415241 and log_256(1/2)=-0.125 */
         if (-0.415241 * ndigits - 0.125 > bytes) {
             return RubyFixnum.zero(runtime);
         }
-        
+
         IRubyObject f = Numeric.int_pow(context, 10, -ndigits);
 
         if (this instanceof RubyFixnum && f instanceof RubyFixnum) {
@@ -443,7 +443,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** integer_to_r
-     * 
+     *
      */
     @JRubyMethod(name = "to_r")
     public IRubyObject to_r(ThreadContext context) {
@@ -457,7 +457,7 @@ public abstract class RubyInteger extends RubyNumeric {
     public IRubyObject rationalize(ThreadContext context, IRubyObject[] args) {
         return to_r(context);
     }
-    
+
 
     @JRubyMethod(name = "odd?")
     public RubyBoolean odd_p(ThreadContext context) {
@@ -483,25 +483,25 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** rb_gcd
-     * 
+     *
      */
     @JRubyMethod(name = "gcd")
     public IRubyObject gcd(ThreadContext context, IRubyObject other) {
         checkInteger(context, other);
         return f_gcd(context, this, RubyRational.intValue(context, other));
-    }    
+    }
 
     /** rb_lcm
-     * 
+     *
      */
     @JRubyMethod(name = "lcm")
     public IRubyObject lcm(ThreadContext context, IRubyObject other) {
         checkInteger(context, other);
         return f_lcm(context, this, RubyRational.intValue(context, other));
-    }    
+    }
 
     /** rb_gcdlcm
-     * 
+     *
      */
     @JRubyMethod(name = "gcdlcm")
     public IRubyObject gcdlcm(ThreadContext context, IRubyObject other) {
@@ -524,11 +524,11 @@ public abstract class RubyInteger extends RubyNumeric {
 
     /*  ================
      *  Singleton Methods
-     *  ================ 
+     *  ================
      */
 
     /** rb_int_induced_from
-     * 
+     *
      */
     @Deprecated
     public static IRubyObject induced_from(ThreadContext context, IRubyObject recv, IRubyObject other) {
